added Lepton.h having common variables between electrons and muons while keeping all the same functionalities
