+++ DataFormats/interface/Electron.h
@@ -3,7 +3,7 @@
 
 #include "DataFormats/Candidate/interface/CandidateFwd.h"
 #include "DataFormats/Candidate/interface/Candidate.h"
-#include "CATTools/DataFormats/interface/Particle.h"
+#include "CATTools/DataFormats/interface/Lepton.h"
 #include "DataFormats/PatCandidates/interface/Electron.h"
 #include "DataFormats/GsfTrackReco/interface/GsfTrack.h"
 
@@ -17,54 +17,29 @@ namespace cat {
 
 namespace cat {
 
-  class Electron : public Particle{
+  class Electron : public Lepton{
   public:
     Electron();
     Electron(const reco::LeafCandidate & aElectron);
     virtual ~Electron();
 
-    float dxy() const { return dxy_; }
-    float dz() const { return dz_; }
+    float electronID(const std::string& name) const;
+    float electronID(const char* name) const { return electronID( std::string(name) );}
 
-    float scEta() const { return scEta_; }
-    bool passConversionVeto() const { return passConversionVeto_; }
-
-    float chargedHadronIso(float dR=0.3) const {
-      if( dR < 0.35) return chargedHadronIso03_;
-      else return chargedHadronIso04_;
-    }
-    float puChargedHadronIso(float dR=0.3) const {
-      if( dR < 0.35) return puChargedHadronIso03_;
-      else return puChargedHadronIso04_;
-    }
-    float neutralHadronIso(float dR=0.3) const {
-      if( dR < 0.35) return neutralHadronIso03_;
-      else return neutralHadronIso04_;
-    }
-    float photonIso(float dR=0.3) const {
-      if( dR < 0.35) return photonIso03_;
-      else return photonIso04_;
-    }
-    float absIso(float dR=0.3, float dBetaFactor=0.5) const{
-      if(dBetaFactor>0 && puChargedHadronIso(dR)<0) return -1;
-      float neutralIso = neutralHadronIso(dR) + photonIso(dR);
-      float corNeutralIso = neutralIso - dBetaFactor * puChargedHadronIso(dR);
-      float charged = chargedHadronIso(dR);
-      return charged + ( corNeutralIso>0 ? corNeutralIso : 0 ) ;
-    }
-    float relIso(float dR=0.3, float dBetaFactor = -1) const {
-      if (dBetaFactor >= 0){
-	float abs = absIso(dR, dBetaFactor)/this->pt();
-	return abs >=0 ? abs : -1;
-      }
+    float relIso(float dR=0.3) const {
       if( dR < 0.35) return relIso03_;
       else return relIso04_;
     }
-
-    bool mcMatched() const { return mcMatched_; }
+    
+    float scEta() const { return scEta_; }
+    bool passConversionVeto() const { return passConversionVeto_; }
+    bool isGsfCtfScPixChargeConsistent() const{ return isGsfCtfScPixChargeConsistent_; }
 
     int snuID() const {return snuID_;}
-
+    
+    void setElectronIDs(const std::vector<pat::Electron::IdPair> & ids) { electronIDs_ = ids; }
+    void setElectronID(pat::Electron::IdPair ids) { electronIDs_.push_back(ids); }
+    
     void setrelIso(double dR, double chIso, double nhIso, double phIso, double AEff, double rhoIso, double ecalpt)
     {
       float relIso = ( chIso + std::max(0.0, nhIso + phIso - rhoIso*AEff) )/ ecalpt;
@@ -73,45 +48,13 @@ namespace cat {
     }
     void setscEta(float i) { scEta_ = i; }
     void setPassConversionVeto(bool i) {  passConversionVeto_ = i; }
-
-    void setChargedHadronIso03(float i) { chargedHadronIso03_ = i; }
-    void setPUChargedHadronIso03(float i) { puChargedHadronIso03_ = i; }
-    void setNeutralHadronIso03(float i) { neutralHadronIso03_ = i; }
-    void setPhotonIso03(float i) { photonIso03_ = i; }
-
-    void setChargedHadronIso04(float i) { chargedHadronIso04_ = i; }
-    void setPUChargedHadronIso04(float i) { puChargedHadronIso04_ = i; }
-    void setNeutralHadronIso04(float i) { neutralHadronIso04_ = i; }
-    void setPhotonIso04(float i) { photonIso04_ = i; }
-
-    void setMCMatched(bool m) { mcMatched_ = m; }
+    void setIsGsfCtfScPixChargeConsistent(bool d) { isGsfCtfScPixChargeConsistent_ = d ; }
 
     void setSNUID(int id) {snuID_ = id;}
 
-    float electronID(const std::string& name) const;
-    float electronID(const char* name) const { return electronID( std::string(name) );}
-    void setElectronIDs(const std::vector<pat::Electron::IdPair> & ids) { electronIDs_ = ids; }
-    void setElectronID(pat::Electron::IdPair ids) { electronIDs_.push_back(ids); }
-
-    void setShiftedEnDown(float f) { shiftedEnDown_ = f;}
-    void setShiftedEnUp(float f) { shiftedEnUp_ = f;}
-    float shiftedEnDown() const {return  shiftedEnDown_;}
-    float shiftedEnUp() const {return  shiftedEnUp_;}
-
-    bool isPF() const{ return isPF_; }
-    void setIsPF(bool d) { isPF_ = d ; }
-    bool isGsfCtfScPixChargeConsistent() const{ return isGsfCtfScPixChargeConsistent_; }
-    void setIsGsfCtfScPixChargeConsistent(bool d) { isGsfCtfScPixChargeConsistent_ = d ; }
-
-    void setDz(float d) { dz_ = d; }
-    void setDxy(float d) { dxy_ = d; }
-
-    bool VetoID() const {return electronID("veto");}
-    bool LooseID() const {return electronID("loose");}
-    bool MediumID() const {return electronID("medium");}
-    bool TightID() const {return electronID("tight");}
-    bool mvaMediumID() const {return electronID("wp90");}
-    bool mvaTightID() const {return electronID("wp80");}
+    bool isVeto() const {return electronID("veto");}
+    bool isMediumMVA() const {return electronID("wp90");}
+    bool isTightMVA() const {return electronID("wp80");}
     
   private:
 
@@ -120,28 +63,9 @@ namespace cat {
     float relIso03_;
     float relIso04_;
 
-    float chargedHadronIso03_;
-    float puChargedHadronIso03_;
-    float neutralHadronIso03_;
-    float photonIso03_;
-
-    float chargedHadronIso04_;
-    float puChargedHadronIso04_;
-    float neutralHadronIso04_;
-    float photonIso04_;
-
     float scEta_;
-    bool isPF_;
-    bool isGsfCtfScPixChargeConsistent_;
-
-    bool mcMatched_;
     bool passConversionVeto_;
-
-    float shiftedEnDown_;
-    float shiftedEnUp_;
-
-    float dz_;
-    float dxy_;
+    bool isGsfCtfScPixChargeConsistent_;
 
     int snuID_;
   };
