+++ DataFormats/src/Lepton.cc
@@ -1,0 +1,14 @@
+#include "CATTools/DataFormats/interface/Lepton.h"
+
+using namespace cat;
+
+/// default constructor
+Lepton::Lepton() {
+}
+
+Lepton::Lepton(const reco::LeafCandidate & aLepton) : Particle( aLepton ) {
+}
+
+/// destructor
+Lepton::~Lepton() {
+}
